+++ nakago_axum/src/auth/claims.rs
@@ -1,9 +1,13 @@
+use std::any::Any;
+
 use async_trait::async_trait;
 use axum::{
     extract::FromRequestParts,
     http::{request::Parts, HeaderValue},
 };
+use biscuit::{ClaimsSet, Empty};
 use hyper::{header::AUTHORIZATION, HeaderMap};
+use serde::{Deserialize, Serialize};
 
 use crate::State;
 
@@ -13,6 +17,49 @@ use super::{
 };
 
 const BEARER: &str = "Bearer ";
+
+/// The Token, including the JWT string and the payload with claims
+#[derive(Default, Debug, PartialEq, Eq, Clone)]
+pub struct Token<T: Default = Empty> {
+    /// The JWT string itself
+    pub jwt: Option<String>,
+
+    /// The decoded JWT claims
+    pub claims: Option<ClaimsSet<T>>,
+}
+
+/// Implement the Axum FromRequestParts trait, allowing the `Subject` to be used as an Axum
+/// extractor.
+#[async_trait]
+impl<T> FromRequestParts<State> for Token<T>
+where
+    T: Default + Clone + Serialize + for<'de> Deserialize<'de> + Send + Sync + Any,
+{
+    type Rejection = Error;
+
+    async fn from_request_parts(
+        parts: &mut Parts,
+        state: &State,
+    ) -> std::result::Result<Self, Self::Rejection> {
+        let validator = state
+            .get::<Validator<T>>()
+            .await
+            .map_err(|_err| MissingValidator)?;
+
+        match jwt_from_header(&parts.headers) {
+            Ok(Some(jwt)) => {
+                let payload = validator.get_payload(jwt)?;
+
+                Ok(Token {
+                    jwt: Some(jwt.to_string()),
+                    claims: Some(payload),
+                })
+            }
+            Ok(None) => Ok(Token::<T>::default()),
+            Err(e) => Err(e),
+        }
+    }
+}
 
 /// The token's Subject claim, which corresponds with the username
 #[derive(Debug, PartialEq, Eq, Clone)]
@@ -35,7 +82,7 @@ impl FromRequestParts<State> for Subject {
 
         match jwt_from_header(&parts.headers) {
             Ok(Some(jwt)) => {
-                let payload = validator.get_payload(jwt)?;
+                let payload: ClaimsSet<Empty> = validator.get_payload(jwt)?;
                 let subject = payload.registered.subject.clone();
 
                 debug!("Successfully verified token with subject: {:?}", subject);
@@ -50,7 +97,7 @@ impl FromRequestParts<State> for Subject {
 
 /// If an authorization header is provided, make sure it's in the expected format, and
 /// return it as a String.
-pub fn jwt_from_header(headers: &HeaderMap<HeaderValue>) -> Result<Option<&str>, Error> {
+fn jwt_from_header(headers: &HeaderMap<HeaderValue>) -> Result<Option<&str>, Error> {
     let header = if let Some(value) = headers.get(AUTHORIZATION) {
         value
     } else {